+++ carts/urls.py
@@ -13,10 +13,10 @@ urlpatterns = [
     path('remove_item/', views.remove_item, name='remove_item'),
     # path('increase_item/', views.increase_item, name='increase_item'),
     # path('dicrease_item/', views.dicrease_item, name='dicrease_item'),
-    path('kakaopay/', views.kakaopay, name='kakaopay'),
-    path('kakaopay/approval/<int:order_id>/', views.kakaopay_approval, name='kakaopay_approval'),
-    path('kakaopay/cancel/', views.kakaopay_cancel, name='kakaopay_cancel'),
-    path('kakaopay/fail/', views.kakaopay_fail, name='kakaopay_fail'),
+    # path('kakaopay/', views.kakaopay, name='kakaopay'),
+    # path('kakaopay/approval/<int:order_id>/', views.kakaopay_approval, name='kakaopay_approval'),
+    # path('kakaopay/cancel/', views.kakaopay_cancel, name='kakaopay_cancel'),
+    # path('kakaopay/fail/', views.kakaopay_fail, name='kakaopay_fail'),
     path('payments/approval/', views.approval, name='approval'),
     path('payments/show_approval/', views.show_approval, name='show_approval'),
     # path('payments/approval/', views.approval, name='approval'),
