+++ carts/views.py
@@ -5,7 +5,7 @@ from django.contrib.auth.decorators import login_required
 from .models import Cart, CartItem, Order, OrderItem
 from stores.models import Product
 from django.http import JsonResponse, HttpResponseNotFound
-import os, requests, json
+import os, requests, json, math
 from dotenv import load_dotenv
 load_dotenv()
 KAKAO_AK = os.getenv('KAKAO_AK')
@@ -132,82 +132,83 @@ def order_page(request):
     return render(request, 'carts/order_page.html', context)
 
 
-def kakaopay(request):
-    order_id = request.POST.get('order_id')
-    order = Order.objects.get(pk=order_id)
-    domain = request.get_host()
-    cnt = 0
-    order_item = 'none'
-    for item in order.order_items.all():
-        cnt += 1
-        if cnt == 1:
-            order_item = item.product.name
-    if cnt > 1:
-        order_item += f' 외 {cnt-1} 건'
+# def kakaopay(request):
+#     order_id = request.POST.get('order_id')
+#     order = Order.objects.get(pk=order_id)
+#     domain = request.get_host()
+#     cnt = 0
+#     order_item = 'none'
+#     for item in order.order_items.all():
+#         cnt += 1
+#         if cnt == 1:
+#             order_item = item.product.name
+#     if cnt > 1:
+#         order_item += f' 외 {cnt-1} 건'
 
-    URL = 'https://kapi.kakao.com/v1/payment/ready'
-    headers = {
-        'Authorization': 'KakaoAK ' + KAKAO_AK,
-    }
-    params = {
-        'cid': 'TC0ONETIME',    # 테스트용 코드
-        'partner_order_id': order_id,     # 주문번호
-        'partner_user_id': request.user.username, # 유저 아이디
-        'item_name': order_item,        # 구매 물품 이름
-        'quantity': cnt,                # 구매 물품 수량
-        'total_amount': order.total(),        # 구매 물품 가격
-        'tax_free_amount': '0',         # 구매 물품 비과세
-        'approval_url': f'http://{domain}/carts/kakaopay/approval/{order_id}/', # 결제 승인시 이동할 url
-        'cancel_url': f'http://{domain}/carts/kakaopay/cancel/', # 결제 취소 시 이동할 url
-        'fail_url': f'http://{domain}/carts/kakaopay/fail/', # 결제 실패 시 이동할 url
-    }
+#     URL = 'https://kapi.kakao.com/v1/payment/ready'
+#     headers = {
+#         'Authorization': 'KakaoAK ' + KAKAO_AK,
+#     }
+#     params = {
+#         'cid': 'TC0ONETIME',    # 테스트용 코드
+#         'partner_order_id': order_id,     # 주문번호
+#         'partner_user_id': request.user.username, # 유저 아이디
+#         'item_name': order_item,        # 구매 물품 이름
+#         'quantity': cnt,                # 구매 물품 수량
+#         'total_amount': order.total(),        # 구매 물품 가격
+#         'tax_free_amount': '0',         # 구매 물품 비과세
+#         'approval_url': f'http://{domain}/carts/kakaopay/approval/{order_id}/', # 결제 승인시 이동할 url
+#         'cancel_url': f'http://{domain}/carts/kakaopay/cancel/', # 결제 취소 시 이동할 url
+#         'fail_url': f'http://{domain}/carts/kakaopay/fail/', # 결제 실패 시 이동할 url
+#     }
 
-    res = requests.post(URL, headers=headers, params=params)
-    request.session['tid'] = res.json()['tid']      # 결제 승인시 사용할 tid를 세션에 저장
-    next_url = res.json()['next_redirect_pc_url']   # 결제 페이지로 넘어갈 url을 저장
-    return redirect(next_url)
-    # return render(request, 'payments/kakaopay.html', context)
-
-
-def kakaopay_approval(request, order_id):
-    url = 'https://kapi.kakao.com/v1/payment/approve'
-    headers = {
-        'Authorization': f'KakaoAK {KAKAO_AK}',
-    }
-    params = {
-        'cid':'TC0ONETIME',
-        'tid': request.session['tid'], #결제 고유 번호
-        'partner_order_id': order_id, #주문 번호
-        'partner_user_id': request.user.username, #유저 아이디
-        'pg_token': request.GET['pg_token'] # 쿼리 스트링으로 받은 pg토큰
-    }
-    res = requests.post(url, headers=headers, params=params)
-    result = res.json()
-    print(dir(result))
-    context = {
-        'result': result,
-    }
-    return render(request, 'payments/kakaopay_approval.html', context)
+#     res = requests.post(URL, headers=headers, params=params)
+#     request.session['tid'] = res.json()['tid']      # 결제 승인시 사용할 tid를 세션에 저장
+#     next_url = res.json()['next_redirect_pc_url']   # 결제 페이지로 넘어갈 url을 저장
+#     return redirect(next_url)
+#     # return render(request, 'payments/kakaopay.html', context)
 
 
-def kakaopay_cancel(request):
-    context = {
-    }
-    return render(request, 'payments/kakaopay_cancel.html', context)
+# def kakaopay_approval(request, order_id):
+#     url = 'https://kapi.kakao.com/v1/payment/approve'
+#     headers = {
+#         'Authorization': f'KakaoAK {KAKAO_AK}',
+#     }
+#     params = {
+#         'cid':'TC0ONETIME',
+#         'tid': request.session['tid'], #결제 고유 번호
+#         'partner_order_id': order_id, #주문 번호
+#         'partner_user_id': request.user.username, #유저 아이디
+#         'pg_token': request.GET['pg_token'] # 쿼리 스트링으로 받은 pg토큰
+#     }
+#     res = requests.post(url, headers=headers, params=params)
+#     result = res.json()
+#     print(dir(result))
+#     context = {
+#         'result': result,
+#     }
+#     return render(request, 'payments/kakaopay_approval.html', context)
 
 
-def kakaopay_fail(request):
-    context = {
-    }
-    return render(request, 'payments/kakaopay_fail.html', context)
+# def kakaopay_cancel(request):
+#     context = {
+#     }
+#     return render(request, 'payments/kakaopay_cancel.html', context)
+
+
+# def kakaopay_fail(request):
+#     context = {
+#     }
+#     return render(request, 'payments/kakaopay_fail.html', context)
 
 def approval(request):
-    print('ApprovalApprovalApprovalApprovalApprovalApprovalApprovalApprovalApprovalApprovalApproval')
     jsonObject = json.loads(request.body)
     print(jsonObject)
     order_id = int(jsonObject['orderId'])
 
     order = Order.objects.get(pk=order_id)
+    order.pay_type = jsonObject['pg']
+    order.postcode = jsonObject['orderPostcode']
     order.address = jsonObject['orderAddress']
     order.phone = jsonObject['orderPhone']
     # order.email = jsonObject['orderEmail']
@@ -215,7 +216,7 @@ def approval(request):
     order.receiver = jsonObject['receiver']
     order.total_price = int(jsonObject['totalAmount'])
     order.use_points = int(jsonObject['usePoints'])
-    order.total_amount = int(jsonObject['totalAmount']) - int(jsonObject['usePoints'])
+    order.total_amount = int(jsonObject['finalAmount'])
 
     order.shipping_status='배송준비중'
     order.save()
